% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_density_adaptive.R
\name{stat_density_adaptive}
\alias{stat_density_adaptive}
\alias{geom_density_adaptive}
\title{Stat for adaptive KDE}
\usage{
stat_density_adaptive(
  mapping = NULL,
  data = NULL,
  geom = "area",
  position = "stack",
  ...,
  bw = "SJ",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  trim = FALSE,
  na.rm = FALSE,
  bounds = c(-Inf, Inf),
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_density_adaptive(
  mapping = NULL,
  data = NULL,
  stat = "density_adaptive",
  position = "identity",
  ...,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  outline.type = "upper"
)
}
\arguments{
\item{bw}{The smoothing bandwidth to be used for the reference density.
If numeric, the standard deviation of the smoothing kernel.
If character, a rule to choose the bandwidth, as listed in
[stats::bw.nrd()].}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible
to adjust the bandwidth while still using the a bandwidth estimator.
For example, `adjust = 1/2` means use half of the default bandwidth.}

\item{kernel}{Kernel. See list of available kernels in [density()].}

\item{n}{number of equally spaced points at which the density is to be
estimated, should be a power of two, see [density()] for
details}

\item{trim}{If `FALSE`, the default, each density is computed on the
full range of the data. If `TRUE`, each density is computed over the
range of that group: this typically means the estimated x values will
not line-up, and hence you won't be able to stack density values.
This parameter only matters if you are displaying multiple densities in
one plot or if you are manually adjusting the scale limits.}

\item{bounds}{Known lower and upper bounds for estimated data. Default
`c(-Inf, Inf)` means that there are no (finite) bounds. If any bound is
finite, boundary effect of default density estimation will be corrected by
reflecting tails outside `bounds` around their closest edge. Data points
outside of bounds are removed with a warning.}
}
\description{
Stat for adaptive KDE
}
